+++ changelog.md
@@ -1,5 +1,20 @@
 # OfficeFileVersionUpdater Changelog
 
+**v1.0.3 [20230805]**
+- NEW & UPDATED:
+  - Rewrote the code to be more OOP. 
+- BUGS & FIXES:
+  - N/A
+
+**v1.0.2 [20230805]**
+- NEW & UPDATED:
+  - Included the Office DLL files. You'll still need Office installed to run but this adds some level of compatibility.
+  - Reworked some of the error/exit handling
+  - Removed unused references (code stuff)
+  - Reformatted the code
+- BUGS & FIXES:
+  - N/A
+
 **v1.0.1 [20230804]**
 - NEW & UPDATED:
   - Changed the command-line argument handling logic. This is largely so that the introduction of further (future) parameters becomes a lot easier to handle.